Don't fail if cannot register preexisting but optional memory segments.
Some preexisting memory segments we have been registering are actually
optional, in that failing to register them would have no functional or
performance impact.  An example is the segment with pathname "[stack]",
which is the stack for the main thread of the process.  In the past the
Chapel main program was hosted by the main process thread and we had to
register that thread's stack, because the main program's stack variables
might be remotely communicated and the early ugni comm layer had no
ability to communicate unregistered memory via bounce buffers.  But some
of these things are no longer true.  We don't run Chapel code on the
main thread any more, and nowadays comm=ugni can bounce unregistered
communication.

In the past there have been unexpected changes with respect to which
preexisting memory segments we could register, and recent pre-release
experience with an upcoming CLE version indicates this will continue.
So here, rewrite the code that registers preexisting segments so that it
is less fragile.  Specifically, attempt to register each segment as we
process it, instead of doing them all at the end, and only report an
internal error and halt if registration fails on a segment we really
expect we should be able to register and which has functional or
performance implications.  Basically this comes down to those with
"huge" anywhere in their pathname.

While here, I also fixed a bug which caused the segments added for
dynamic allocations prior to processing /proc/self/maps to be registered
twice, apparently to no ill effect.

Also while here, remove the logic charged with producing reasonable
behavior when there were more registered segment candidates than we had
table entries.  Table entries are plentiful now, so running out can just
be an internal error.