Update and dramatically simplify this test for no enum->int coercions
The maintenance cost for updating this test now that there are no enum->int
coercions outweighs its value.  Since the original point of the test was
to note that you could have odd behavior if you had two enums with the
same name, I reduced the test down to something that demonstrates that
yet is not difficult to maintain.  I also added a test for the error
cases.