Be sure to compute max comm domains before needing it!
Recent (last several months) code movement led to a situation in which,
for minimal-registered-memory mode, we used the job-wide maximum number
of comm domains per node before we had computed that.  Here, move the
computation and jobwide distribution of that value prior to the point at
which we need it.

Subsequent testing revealed an additional bug: we have been using the
number of logical processors when estimating how many comm domains we
will need.  Using the tasking layer's max level of parallelism is much
better.

That in turn pointed out yet one more bug, less serious but still noisy:
the check to make sure the pool of remote fork completion flags was big
enough was comparing the expected maximum demand for flags to the number
of flags per pool, not the total number of flags in all pools.  So here,
fix that as well.