Update enum/ferguson tests now that enums don't coerce to ints
For two of these tests, I updated a bit to make sure we were
getting expected behavior.  The third felt redundant with others
that now exist and as though it was testing something that no
longer has an equivalent so I retired it rather than trying to
morph it into something new.