Process /proc/cpuinfo for CPU counts in a better way.
We've been computing node core counts by dividing the number of
'processor' entries in /proc/cpuinfo by the number of hardware threads
per core.  We've been getting the latter two values by looking at the
'siblings' and 'cpu cores' entries, and to avoid being confused we've
insisted that all the 'siblings' entries have the same value and all the
'cpu cores' entries have the same value.  But we've begun to encounter
situations where that doesn't hold, which leads to an internal error and
we cannot continue.

Here, switch to just counting the number of distinct pairs of 'physical
id' and 'core id' values to give us a core count.  For at least the one
internal error situation for which we have a copy of the /proc/cpuinfo
file, this gives the correct result.

While here, I also did a little logic rearranging that I think makes
things easier to read.  Specifically, the case where the /proc/cpuinfo
parsing gave us the information we need entirely precedes handling when
it didn't, instead of sort of being split across that.