Simplify coding.
Instead of an if-then-else block with a really big else-part to handle
the case where /proc/cpuinfo didn't give us enough information, do an
early return in the then-part.  This allows us to pull the else-part's
body out to the main level, making everything easier to read.  This
change is easier to look at with a diff that ignores leading white space
differences.