Fix bug in MatrixMarket
The call to open() was passing an enum to a place where an int was
expected and happened to get unbuffered behavior as a result.  This
fix, suggested by Michael Ferguson, inserts a flush at a key place to
avoid the need for unbuffered I/O and resolves the argument mismatch
by removing it.