Remove reliance on enum->int coercions
Here, I've changed an arithmetic array to an array over enums in
order to simplify most of the indexing within the program.
There are some casts back from ints to enums required due to the
fact that we don't have enum ranges yet.  Once we do, this code
should be revisited.