Convert ref(unmanaged) to canonical as well
Addresses problems with distributions primer.

The reasoning here is that
 * reference types are special (refs added/removed during passes)
 * a ref(borrow) is convertible to a ref(unmanaged) since
   these have the same management

These two factors aren't true for say record types - there it
would be a type error to (say) use a pointer to a record with
a managed field when a pointer to a record with an unmanaged field
is needed.