Reflect review feedback: enlarge cpuId table, improve default numPUs.
The kernel seems to allow for up to 8192 PUs, so make the table that
size.  We expect to size it dynamically in the future, but this will do
as a stopgap.  Also, instead of a simple '1', use the number of unique
physical and core IDs as the default number of processors if we don't
see any 'processor' lines in /proc/cpuinfo.