add printInitCommCounts.lm-numa.good
The only failure in a full paratest under numa and comm=gasnet,
as of this writing, is in this test:

  modules/sungeun/init/printInitCommCounts.chpl

because of higher comm counts compared to printInitCommCounts.good.
The increase is that #gets on non-0 locales is 25, vs. 14 for flat.

We do not track this test in the nightly testing and do not have plans
to address this increase at the moment.

Therefore I am "locking in" these higher comm counts for numa
by adding a numa-specific .good. That way a full paratest over numa
is back to clean.