Update chpl_arch.py to handle 64-bit ARM, KNL, and Skylake (#9013)
Update chpl_arch.py to handle 64-bit ARM, KNL, and Skylake (#9013)

[reviewed by @ronawho and @ben-albrecht ]

ARM presents a challenge for chpl_arch.py because Intel architectures are accessed with the -march= compiler flag, and ARM architectures are accessed with the -mcpu= flag. This change includes an indication of which flag to use in the output of chpl_arch.py and updates it to handle the relevant ARM architectures.

This allowed the removal of some special cases from clangUtil.cpp.

Also, KNL and Skylake have been added.

Makefiles had to be changed to expect CHPL_MAKE_TARGET_BACKEND_ARCH to include part of the flag. Some Python scripts were changed to stop after the first = when parsing environment variables.

This PR addresses issues #5686 and #8820 .