Minor improvements to paratest.server
* When aggregating individual log files, do so in a sorted order.

This is to make the order of tests in the combined log consistent
among multiple runs. In the spirit of #8572, this order will not
necessarily be consistent across systems, users, etc.

Note that the memleaks logs are already aggregated in a sorted order.

* At the end of a paratest run, print information messages that
the logs are being aggregated ("collected"). This reminds the user
where the final log(s) will be.