Fix gasnet build by removing GASNETI_FALLTHROUGH
In 6eebb6215a I mistakenly brought in an upstream patch that I thought was just
a comment update so that I could cleanly bring in bc9313d389. Turns out
GASNETI_FALLTHROUGH isn't defined in the release of gasnet that we're using
only upstream. Remove GASNETI_FALLTHROUGH to avoid build issues.