Extend performance testing to allow a test to only have a .perfcompopts
We used to only recognize a performance test if it had at least .perfkeys or a
.perfexecopts file, but failed to recognize it if we only had a .perfcompopts
file.  Since .perfcompopts files can specify different .perfkeys, we should
allow tests to only specify a .perfcompopts if they so choose.

With this change, miniMD will run when called from the directory (with the
changes in the previous commit)