Revert #9205 "Workaround for c_sublocid_any ..."
Reverts #9205 "Workaround for c_sublocid_any -related failures"

The workaround was to avoid the following issue.

An internal module contained a local variable whose name was the same
as the name of a global 'extern' variable. The compiler failed to
uniquifyName() of the local variable because it did not realize that
it was necessary. The latter happened because the extern variable
was !isRenameable() so its name was not recorded. The local variable's
name, then, was not deemed to be in conflict with the extern.

#9300 fixed this behavior, so the workaround from #9205 is no longer
necessary.