Add two tests of applying .type to a generic arg -- one working, one future
genericArgDotType.chpl shows that when an argument's type is based on
that of an argument with a specified, but formal, generic type, the
compiler only inherits the loose generic type of the argument, not the
more specific actual type with which it was instantiated.  This seems
surprising and wrong.

In contrast, genericArgDotType2.chpl shows that when the generic
argument simply has no type specified, the type constraint is taken
based on the actual type.  I believe that the first case should work
like this as well.