Update tuples to use special function instead of constructors
We aim to eventually deprecate constructors, so this commit updates
tuples to use a special function instead of a constructor. Tuples are
sufficiently special that I don't think it makes sense to use an
initializer just yet.

This special function is named 'chpl__init_tuple', and is generated by
the compiler in the same way the old constructor was generated. Various
conditionals in the compiler have been updated to look for FLAG_TUPLE
instead of a function name.