Adjust getInstantiationType & checkGenericFormals
The compiler already has code to select the appropriate
generic type to instantiate when the generic type requires
some form of coercion to get to it.

This commit cleans up this code in the compiler and updates
it to also apply to sync/single variables when passed to a
generic argument with a generic type that applies to the contents
of the sync/single.

Consider the following example:

  proc f(x:enumerated) {
    writeln(x);
  }

  enum e {a=1};

  var s:sync e = e.a;
  f(s);