Move == and != on enums to ChapelBase.chpl
In his code review, Michael pointed out that there were enumerated
overloads of comparison operators in ChapelBase.chpl still, and
this showed that we were still supporting comparisons between
mixed enums in a param context.  This also showed that we could
support comparisons on enums by defining them in module code rather
than in buildDfeaultFunctions.cpp which seems attractive and
arguably more consistent with what we do for other scalar types.

This is a prototype of making that switch where buildDefaultFunctions.cpp
still needs to be cleaned up before this could be merged; and also
we may want to look into moving the enum assignment out into module
code as well (if not all enum routines?)