arithmetic operations on Series ADT (#9342)
[reviewed by @benharsh]

The uni() method combines the indexes of both Series while taking a SeriesUnifier object, which contains three overridable functions: f(lhs, rhs), f_lhs(lhs), and f_rhs(rhs). These three functions are used to fold the values of the lhs and rhs Series, respectively.

add(), subtr(), and mult() utilize uni(), and operators (+, -, and *) are built on top of those functions in turn. (The nested implementation of uni() and the other functions is necessary for dynamic dispatch to realize the types of the Series and the Index.)