Make the thread index a first-class datum.
We've long supported the notion of a thread index, but only for adding
context to the debug output.  Here, make it available always, not just
for debug support.  We will need it in coming work to reduce contention
in the the remote-fork-done pools.

While here, also bump the thread index size to 64 bits from 32, as a
hedge (and because in our foreseen usage it won't hurt performance).