Implement <=> for 'Shared'
This implements the swap operator for 'Shared' just as PR #9343 did
for Owned.  This isn't strictly necessary, as the default swap
implementation worked; but this results in a more efficient and native
implementation.

Extended my previous swap-owned.chpl file to include shared objects
and renamed it swap-managed.chpl.

While here, I realized that I'd dropped the swap operator into a
fairly arbitrary place in OwnedObject, and moved it closer to the
assignment operator overloads since they're thematically related.