Improve scopeResolve for ShadowVarSymbol
Before this change, scope resolution for the outer variable "ovar"
of a shadow variable "svar" was done so:

* the parser reprsented the ovar as UnresolvedSymExpr
* scopeResolve handled that UnresolvedSymExpr just like any other
* ... and converted it to SymExpr

One outcome of this was that ShadowVarSymbol::outerVarRep could
not have the type SymExpr*. This was not ideal because it was
actually a SymExpr* for most of the compiler.

Now, scope resolution works like this:

* the parser does not represent the ovar
* scopeResolve creates the ovar SymExpr for each svar
  using the same steps as if it were an UnresolvedSymExpr,
  except it uses the svar's name instead of
  UnresolvedSymExpr::unresolved

As a result, there is no UnresolvedSymExpr for the ovar
and ShadowVarSymbol::outerVarSE has the type SymExpr* .
outerVarSE replaces outerVarRep.

Also, when there is no outer variable for a given name,
we give a specific error message, rather than a generic one.
Detecting the case of the shadow variable referring to
a field or a function (for error reporting) is also easier now.

The tests for the error messages still do the same as before.
The changes are:

* update the error message text
* split a test into two and otherwise separate the checking
  of with-clauses for foralls from those for task constructs (*)

(*) I had to introduce an earlier USR_STOP(). Because there is
another (existing) check for err_user (or somesuch) that is
interpreted as "the immediately-preceeding operation failed
and we can recover". So some errors that used to be reported
in a single compiler run no longer are - we need to test them
separately.