Enable auto-copy of owned
Because such auto-copies can introduce const checking errors,
run late const checking *again* in callDestructors.

Late const checking already runs in cullOverReferences and
makes use of a data structure local to that pass to print out
the reason something was const or not. An alternative to running
it twice would be to make this data structure global. But, it might
be out of date by the it runs in callDestructors, so I think we'll
get better errors if we run it twice.