check_paths: Improve argparse usage
Provide an example invocation including a shell loop, to tide us over
until the script learns to walk (directories).  The particular length
of the descriptive part of the string gets argparse to put the
example shell code on its own line.

In that same meantime, make the -f argument required.