filtering on Series ADT (#9371)
[reviewed by @benharsh]

The map() method iterates over a Series while taking a SeriesMapper object, which contains the overridable function f(d). This function is used to create a new Series with the same Index, but different values. All the comparison operators are implemented with this pattern via SeriesCompareScalar, emitting a TypedSeries(bool) based on f(d).

The filter() method takes in a TypedSeries(bool) and emits a subset of 'this' Series. The subset is defined as follows:

- The passed Series assigns true to a given index.
- 'this' series contains that index.

In turn, filter() is used by TypedSeries.this(). Taken together, map() and filter() enable the following elegant pattern:

var A = TypedSeries(...);
var ltThree = A[A < 3];