Fix locale_name.prediff
I was getting comm and substrate values from the env vars, but we don't stuff
the chplenv into prediffs/precomps/preexecs like we do for other files, so the
CHPL_COMM_SUBSTRATE wasn't set on gasnet-darwin testing. Other configs use
paratest, and paratest.server ends up propagating and setting the chplenv so
this failure mode was masked in other configurations (and my testing)

This just updates the prediff to explicitly run the chplenv scripts instead of
relying on the env vars.