Add a test of a 'type' argument with a constraint and a query expression
While answering a user question last week, I was excited to find that
the following case worked as I'd hoped: putting a generic type
constraint on a `type` argument supported the ability to do queries of
the generic elements of the type, as in value argument contexts.
Doing some grepping I wasn't easily finding instances of other tests
that tried to exercise this pattern so thought I'd add this to lock
the behavior in or put more weight on it.