Use different code for makeRawNew in resolve
The makeRawNew function was called both from build and
from implementForallIntents during resolution. The problem
with that is that the AST for new is slightly different before
and after normalize. So, solve the problem by moving the function
to build.cpp and getting it to generate the pre-normalize pattern
and then using the body the function in implementForallIntents with
the post-normalize pattern.