Make mp_limb_t opaque and adjust getlimbn to match
Because GMP can be configured to create a different mp_limb_t
on different systems (and in fact it does so on Cygwin), the
GMP module should not assume anything about the size of mp_limb_t.
This commit makes that type opaque. Some related code needed
to be adjusted based on that change:

 * chpl_gmp_get_mpz now uses chpl_comm_get and sizeof
   instead of chpl_comm_array_get
 * added chpl_gmp_mpz_nlimbs and chpl_gmp_mpz_getlimbn. These are
   callable from user code & used in the BigInteger module to
   get the number of limbs / the i'th limb. These both return uint(64)
   always, so that the issue with mp_size_t varying is less problematic.
 * BigInteger functions numLimbs and get_limbn now use the uint(64)
   versions above.
 * mpz_getlimbn is now private to GMP.chpl since it can't be
   used right now without internal compiler knowledge, since the
   Chapel compiler won't understand its return type.

While adjusting numLimbs and get_limbn I noticed that these do not
appear to be tested, so I added a (multilocale even!) test that
they work correctly.