'None' entries in Series ADT (#9429)
[reviewed by @benharsh]

A valid_bits array has been added to the TypedSeries class, along with internal iterators emitting those bits. These iterators were then utilized by the union, map, and filter operations to properly process the underlying data. These operations are still carried out on the underlying data as if it were valid, however it is masked with the valid_bits array in the final result.

- For union, both datum have to be valid for the output to be valid.
- For map and filter, the original datum has to be valid for the output to be valid.