Allow specifying /proc/cpuinfo system file name via the environment.
On Linux-based systems the runtime looks at /proc/cpuinfo to help figure
out how many processors and hardware threads are present.  Here, allow
the name of this file to be specified via the environment.  This is a
maintenance aid, so that we can feed outside users' /proc/cpuinfo files
into the runtime when debugging problems.