Add tests for referencing a global variable from an exported function
They demonstrate that if the exported module's init function is not called,
global variables defined in it will not be given their proper value and so
calls to functions that reference them will not behave as expected.  They also
demonstrate that we cannot call the module init function ourselves just yet.