Preserve runtime type for formals with 'in' intent
This commit updates wrapper generation to preserve the runtime type of
formals with the 'in' intent. For example:

```chpl
proc foo(in D : domain(1), in A : [D] int);
var X : [1..4] int;
foo({1..4}, X);
```

The formal 'A' is declared to have the runtime type '[D] int'. Anything
else would be surprising/confusing. Prior to this commit we would simply
initCopy X and pass the result to 'foo', which would have a different
runtime type than '[D] int'. To correct this, we can instead create a
temporary for 'A' at the callsite and assign 'X' into that temporary:

```chpl
var default_arg_A : [actual_D] int;
default_arg_A = X;
foo(actual_D, default_arg_A);
```

This commit leverages the pre-existing default_arg creation
functionality to create the temporary into which X is assigned.
Generally, with this commit formal temps are created as if they were
variable declarations.

In order to support this new functionality, FLAG_DEFAULT_ACTUAL is
applied to a certain temporary and later used to avoid inserting an
initCopy. FLAG_NO_AUTO_DESTROY is also applied to default actuals passed
by in-intent.