Modify default initializers to use 'in' intent
Using the 'in' intent for default initializers achieves the desired
behavior when passing formals or relying on default values.

Instead of initializing a field and assigning a formal into it, default
initializers will now 'move' the formal into the field.

Resolution has been updated to allow handling of default initializers to
use the common path, instead of the path for constructors. Wrapper
handling is updated to avoid creating extra copies for _new wrappers or
super.init calls within a default initializer.