Improve non-kernel scalability of isx-hand-optimized
The timed kernel of ISx was scaling well, but the total running time was
scaling horrendously. The root cause of this was slow destruction of some
arrays. We have a bunch of block distributed arrays with inner [1..numTrials]
arrays for timer values. During deinit, all remote locales were going to locale
0 to remove themselves from the list of arrays declared over `1..numTrials`.
This led to a ton of on-stmts, each trying to acquire a shared atomic-lock
which killed scalability.

As a hacky workaround this adds a simple record wrapper so that we end up
creating unique `1..numTrials` domains instead of there being a single one.
This significantly improves scalability of ISx's total running time.

| nodes | kernel time | total before | total now |
| ----- | ----------- | ------------ | --------- |
|  64   |  9.6        |  24.6        |  19.6     |
| 128   | 10.4        |  40.1        |  22.1     |
| 256   | 11.6        |  95.5        |  25.5     |
| 512   | 12.8        | 218.1        |  28.1     |