Fix miscellaneous bugs in Reindex bulk transfers
An array reindex was incorrectly translating indices from the upward
domain to the downward domain. This was caused by passing the wrong
upward domain class to chpl_reindexConvertDom, and is fixed by passing
'privDom.updom'.

After this bug was fixed, it became apparent that we were also
incorrectly translating strided slices of reindexes. This is solved by
cloning chpl_reindexConvertDom and adjusting it to preserve the slice's
stride.