Work around disappearing instantiation points for initializers
Sometimes we create a BlockStmt, insert calls to be resolved, and then
remove or flatten that BlockStmt. Any new functions resolved within that
block will then be pointing to a dead AST node.

We used to track the instantiation of an AggregateType by using the
initializer's instantiation point. This commit attempts to work around
the disappearing instantiation point by instead using the body of the
function that instantiated the initializer.