Remove a bad `assert(0)` from qio's setup_actions
Ben A and I found a bug when spawning a task while setting stdout to PIPE and
stderr to STDOUT.  If the runtime was compiled without optimizations (if
CHPL_DEVELOPER is true), we would encounter an assertion error.  Basically, we
don't want to insert a close action if the file descriptor provided is STDOUT,
but the branch also had an `assert(0)` and that caused the program to halt.
Michael says it is fine to just remove the assert, and behavior seems to be
correct without it.