deprecate 'out error' pattern from modules (#9450)
[reviewed by @mppf]

The out error: syserr pattern was initially created for the IO module, and became the de facto error handling approach within Chapel modules. Now that a fleshed out approach is in place, the standard modules ought to use it. Additionally, the standard modules provide a solid proving ground for the error handling feature.

Implementation:
- Move functionality from the out error: syserr function to the throws function.
- Rewrite the out error: syserr function as a wrapper for the throws function.
- Add a deprecation warning to the out error: syserr function.