Valgrind documentation, with re2 support (#9561)
[reviewed by @ronawho, @benharsh]

The CHPL_RE2_VALGRIND_SUPPORT compiler flag was not well documented. This PR creates Valgrind.rst to document both the compiler flag and other environment variables needed for chpl generated binaries to work with valgrind.

This PR also edits start_test to require that CHPL_RE2_VALGRIND_SUPPORT is set if the valgrind and CHPL_REGEXP flags are set.