Tighten up resolveCoerce
resolveCoerce adjusts PRIM_COERCE in a special manner
to handle the return type for tuples that are returned.
But this adjustment does not make sense for variable initialization,
and now PRIM_COERCE is used in variable initialization. So limit
the adjustment to the case in which the result of PRIM_COERCE is returned.