Remove BlockCycDist's copy initializer
The copy initializer wasn't a true copy initializer, since it had an extra
argument for the type and param fields.  Those aren't necessary in the
initializer world, and with the arguments removed, the initializer looks
basically identical to what the compiler generated copy initializer would look
like, so remove it entirely.