Fix reversed sense of vectorize enable flag (#9577)
Fix reversed sense of vectorize enable flag (#9577)

[merged to fix breakage in master, asking @mppf to do a post-commit review]

In #9567 , explicit vectorization metadata was enabled when not using the Region Vectorizer, causing messages about failed vectorization to print out. This causes some tests to fail.

This happened because the metadata was enabled when not using the RV, and disabled when using the RV. I believe it was meant to be the other way around.