Tolerate ABI variations in LLVM calling sequence (#9585)
Tolerate ABI variations in LLVM calling sequence (#9585)

[reviewed by @mppf ]

The 64-bit ARM calling sequence ABI is a little different from x86_64. A struct with a size less than 128 bits is rounded up to a multiple of 64 bits. This change tolerates that difference so that ARM code can work.

This change brings us from zero working ARM tests to all of release/examples working except for the ra benchmarks. Because testing shows that x86_64 has not been broken in the process, and because this is an improvement, I am preserving this change by committing it even though not all of the ARM tests work yet.

The ra benchmarks on ARM get an internal LLVM assertion failure from "extracting off end of register," which means I must be missing a special case. The assertion failure does not occur until the makeBinaryLLVM() stage, so it will take a while to debug.

A late-breaking development: the ra benchmarks only fail in multilocale mode, which may provide a clue.