Use RDMA for puts larger than RDMA_THRESHOLD under ugni
RDMA_THRESHOLD is currently defined as 1024 bytes based on what cray-mpich
does, but it's probably worth doing some tuning and allowing it to be
overridden with an environment variable.

Note that RDMA operations require the local side to be registered as well, so
we have to check the size and local registration.