Update TransformLogicalShortCircuit to lower to IfExprs
If-functions are no longer handled, so lower to IfExprs instead.

This commit also updates the transformation to occur pre-order and to
always insert the new IfExpr before the operator's containing statement.
This allows for IfExprs to be nested correctly, and preserves the old
behavior because the outermost IfExpr is inserted before its containing
parent.