Lower IfExprs to CondStmts during normalization
This commit introduces a visitor-transformation that turns IfExprs into
CondStmts. This transformation occurs in pre-order so that nested
IfExprs turn into nested CondStmts.

The resulting CondStmt must always end in a PRIM_MOVE where the
left-hand side is a symbol with FLAG_IF_EXPR_RESULT and the right-hand
side is a PRIM_LOGICAL_FOLDER. These properties will later be leveraged
by resolution in order to correctly infer the 'return' type of the
if-expr.

This commit also makes minor changes to variable initialization in order
to handle IfExprs.