Update resolution to handle CondStmts lowered from IfExprs
An IfExpr's branches may return different types provided that one type
can coerce into another. Resolution would normally infer the result's
type by looking at the RHS of a PRIM_MOVE, but here we want to wait
until the entire CondStmt is resolved. This means we need to skip some
portion of PRIM_MOVE-resolution for symbols with FLAG_IF_EXPR_RESULT.

Once we have resolved a CondStmt, and that statement moves into an
IF_EXPR_RESULT, we call the new function resolveIfExprType.

This function is similar to 'resolveReturnTypeAndYieldedType' and
leverages the same functionality. If the returned type from each branch
is not the same, then one type must be coerce-able into the other.

This function handles the following error cases:
- Returning incompatible types
- Returning a type from one branch, but a value from the other
- Returning types from both branches (suggests using param-conditional)

This function is also responsible for inserting copies should one branch
return a reference and the other return a value.