Add a test from unmerged PR #4308
This test illustrates an issue with the compiler, where in the code:

```chpl
class MyType {}
proc test(arg:MyType) {}
test(nil);
```

where MyType is otherwise unused, e.g. no instances or subclasses of
it are created, MyType is pruned during resolution.

#4308 offered a solution, which was not deemed appropriate.
See that PR for more details.
For now, just adding the test.