Add boilerplate and some content for domain
Adds some notes for myself in the _getArray implementation.  Adds the start of
the domain class, including some fields and some iterators.  Starts on
dsiBuildArray but I need to think more deeply about what I want to do in that
situation, since the instance field in the array class is a const c_ptr and I
can't overwrite that after creation - initCopy assumes I can transfer the
contents into the array after I create it but that would require my pointer
looking at valid memory to start and if it isn't the memory I got originally
then I will be copying stuff over and . . .

Lots to think about there, but I suspect I need to sleep on it