Initial support for enum ranges
This adds very limited support for enum ranges and the start of a
test to lock the capabilities in.  Most of the work here is
adjusting ranges to be able to use a distinct internal type for
representing bounds and strides than those that will be visible
back to the user at the interfaces.