Improve type query syntax for arguments
 * removes PRIM_TUPLE_AND_EXPAND
 * handles nested types e.g. proc f(arg:MyRecord(MyClass(?t)))
 * querying type for varargs indicates all varargs argument have same
   type

The main tricky part here is thas PRIM_QUERY called within PRIM_QUERY
needs to return a value rather than a type (because we don't necessarily
have type accessors for all param/type/untyped fields and because of
runtime types). This presented some challenge in the current structure
of the compiler and I handled it by adjusting preFold to handle this case.
I also considered adjusting the normalizer to put a different flag on
the temps for the outermost PRIM_QUERY or adding a primitive to convert
to a param or a type.